+++ changelog.md
@@ -1,1 +1,2 @@
 - Added bamboo, cherry and mangrove planes
+- Added planes creative tab